Added support for financial securites as pseudo-currencies
tl;dr: Made it possible to refer to financial securities as currencies by
prefixing their symbols with "^". For example, creating an account that uses
"^AAPL" as its currency will make it refer to shares of Apple Computer's
publicly traded stock on NASDAQ. Note: There's not yet any GUI that can be
used to create security pseudo-currencies, they have to be added manually by
editing the .moneyguru file (or programatically with your own code).

* Added a new yahoo_financial_provider plugin that will query the Yahoo
  Finance historical database of stock quotes [using the external
  yahoo_finance module].

* Added support for loading and saving user defined currencies in .moneyguru
  files using the format:

  <currency code="<code>" name="<name>" [exponent="<decimals>"]
   [base="<base-currency-code>"] />

  The base currency is only meant to be used with financial securities and
  define what [real] currency that security is traded in, e.g:

  <currency code="^AAPL" name="Apple Computer" exponent="4" base="USD" />

* Made the currency rate [ex]changer handle security pseudo-currencies as
  per the above definition. Note that cached security prices will be saved
  in the security's own base currency, not CAD.